requirements.txt
# ARISS Requirements
# Install with: pip install -r requirements.txt
#
# NOTE: If you encounter issues with anthropic package,
# try: pip install anthropic --upgrade

# Core dependencies
anthropic
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Web scraping and APIs
praw>=7.7.0                    # Reddit API
google-api-python-client>=2.100.0  # YouTube API
tweepy>=4.14.0                 # Twitter/X API
beautifulsoup4>=4.12.0
requests>=2.31.0

# Sentiment analysis
textblob>=0.17.1
vaderSentiment>=3.3.2

# Database
sqlalchemy>=2.0.0

# Web interface
streamlit>=1.30.0
plotly>=5.18.0

# NLP utilities
nltk>=3.8.0
